Adding initial Guardian authentication to the UserSocket. Now all channels start authenticated through the JWT.
Making it possible to send direct messages to a user through a uniquely identified channel, so private messages can be exchanged. Both from the Socket and through API.

Also adding an API endpoint /api/organizations/:slug/applications/:app_key/event/:event to be able to send messages out of the socket, from an outside source.

The producer external app must have the UserToken associated with the owner of the Application to sign the JWT and this JWT can be used by both the Socket and API authentication.

TODO: add specs to the API and Channel